Move AithUtil
